+++ src/ocean/util/container/map/Map.d
@@ -160,6 +160,29 @@ public class StandardKeyHashingMap ( V, K ) : Map!(V, K)
 
     /***************************************************************************
 
+        Constructor.
+
+        Params:
+            allocator = custom bucket elements allocator
+            n = expected number of elements in mapping
+            load_factor = ratio of n to the number of internal buckets. The
+                desired (approximate) number of elements per bucket. For
+                example, 0.5 sets the number of buckets to double n; for 2 the
+                number of buckets is the half of n. load_factor must be greater
+                than 0. The load factor is basically a trade-off between memory
+                usage (number of buckets) and search time (number of elements
+                per bucket).
+
+    ***************************************************************************/
+
+    public this ( IAllocator allocator, size_t n, float load_factor = 0.75 )
+    {
+        super(allocator, n, load_factor);
+    }
+
+
+    /***************************************************************************
+
         Mixin of the toHash() method which is declared abstract in BucketSet.
 
     ***************************************************************************/